pyproject.toml
[project]
name = "semantic-network-formation-sheaves"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "einops>=0.8.1",
    "gdown>=5.2.0",
    "hydra-core>=1.3.2",
    "igraph>=0.11.9",
    "jax>=0.7.2",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "omegaconf>=2.3.0",
    "polars>=1.33.1",
    "pyarrow>=21.0.0",
    "pytorch-lightning>=2.5.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "timm>=1.0.20",
    "torch>=2.8.0",
    "wandb>=0.22.1",
]

[tool.ruff]
target-version = "py310"
line-length = 79

[tool.ruff.lint]
extend-select = [
      "C4",
      "SIM",
      "UP",
      "TCH",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
